refactor: components/sections 폴더 추가 및 section 컴포넌트 이동 #22
