+++ src/components/sections/ContentSection/ContentSwiper.jsx
@@ -3,9 +3,9 @@ import 'swiper/css/navigation';
 
 import { Swiper, SwiperSlide } from 'swiper/react';
 
-import Arrow from '../../assets/arrow.svg';
-import ContentCard from '../common/ContentCard';
-import ContentCardSkeleton from '../Skeleton/ContentCardSkeleton';
+import Arrow from '../../../assets/arrow.svg';
+import ContentCard from '../../common/ContentCard';
+import ContentCardSkeleton from '../../Skeleton/ContentCardSkeleton';
 import { Navigation } from 'swiper/modules';
 
 const ContentSwiper = ({ data, state, refs }) => {
